+++ src/SDLGameObject.cpp
@@ -11,7 +11,7 @@
 #include "InputHandler.h"
 
 SDLGameObject::SDLGameObject(const LoaderParams* pParams) :
-GameObject(pParams), m_position(pParams->getX(), pParams->getY()), m_velocity(0, 0), m_acceleration(0, 0)
+GameObject(pParams), m_position((float)pParams->getX(), (float)pParams->getY()), m_velocity(0, 0), m_acceleration(0, 0)
 {
     m_width = pParams->getWidth();
     m_height = pParams->getHeight();
