+++ src/main.cpp
@@ -6,7 +6,7 @@
 
 //fixed FPS
 const int FPS = 60;
-const int DELAY_TIME = 1000.0f / FPS;
+const int DELAY_TIME = (int) (1000.0f / FPS);
 
 //main function - initialize the game
 int main(int argc, char *argv[]) {
